training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Package for training toolkit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autoflake>=2.3.1",
    "black>=25.9.0",
    "easyocr>=1.7.2",
    "ipykernel>=6.30.1",
    "isort>=6.1.0",
    "matplotlib>=3.10.6",
    "onnx>=1.19.0",
    "onnxruntime>=1.23.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "ruff>=0.13.3",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
